requirements.txt
cryptography==44.0.2
python-docx==1.1.2
jproperties==2.1.2
gitpython==3.1.44
docx2pdf==0.1.8
openpyxl==3.2.0b1
requests~=2.32.3
dropbox==12.0.2
pandas==2.2.3
pillow==11.1.0
numpy==2.2.4
pytz==2025.2
lxml==5.3.1
six==1.17.0
docx~=0.2.4
